+++ CHANGELOG.md
@@ -26,6 +26,7 @@
         - Allow multiple thread groups to be select.
         - Add continue/interrupt for selected thread group(s).
         - Add run/start for selected thread group(s). Disabled for now.
+* Changing the Qt Style can be saved using Settings->Save Configuration.
 
 ## [1.11] - 2022-09-26